Script, met data updates new Folsom+ResSim model alt
Fixes to scripting, including forecast scripts to tributaries.  Cloud and irradnaicne data is now downloading from DMS, so it was incorporated into forecast scripting. A new model alternative was created that uses W2 for folsom, and the rest uses a new ResSim alternative that included Lake Natoma and the American River.